ExpenseServiceImplTest complete; migrated from BigDecimal to Double in DTOs;

+++ app-server/src/main/kotlin/com/github/arhor/simple/expense/tracker/service/mapping/ExpenseMapper.kt
@@ -14,5 +14,5 @@ interface ExpenseMapper {
     @Mapping(target = "id", ignore = true)
     fun mapToEntity(dto: ExpenseRequestDTO, userId: Long): Expense
 
-    fun mapToDTO(entity: Expense, total: BigDecimal): ExpenseResponseDTO
+    fun mapToDTO(entity: Expense, total: Double): ExpenseResponseDTO
 }
